requirements.txt
certifi==2020.12.5
chardet==4.0.0
greenlet==1.0.0
idna==2.10
numpy==1.20.1
pandas==1.2.3
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
schedule==1.0.0
six==1.15.0
SQLAlchemy==1.4.2
urllib3==1.26.4
pytimedinput==1.3.1